Code re-organization and splitting dev script build
