+++ package.json
@@ -10,6 +10,7 @@
         "build:prod": "npm run build:clean && webpack --env.prod",
         "build:package": "npm run build:clean && webpack --env.prod --env.package",
         "build:clean": "rimraf ./extension",
+        "build:deptree": "depcruise --exclude '^node_modules' --output-type dot ./src --max-depth 10 | dot -T svg > /tmp/dependencygraph.svg",
         "watch": "npm run build:clean && webpack --watch",
         "watch:notif": "webpack --watch --env.notifs",
         "cache:clean": "rimraf ./node_modules/.cache",
@@ -151,6 +152,7 @@
         "css-loader": "^0.28.11",
         "datauri": "^1.1.0",
         "deep-object-diff": "^1.1.0",
+        "dependency-cruiser": "^8.0.0",
         "dotenv-webpack": "^1.7.0",
         "eslint": "^4.19.1",
         "eslint-config-prettier": "^2.9.0",
@@ -180,6 +182,7 @@
         "jest-sourcemaps": "^1.0.1",
         "jest-webextension-mock": "^3.2.0",
         "js-combinatorics": "^0.5.4",
+        "jsdom": "^16.2.0",
         "material-design-icons": "^3.0.1",
         "mini-css-extract-plugin": "^0.4.0",
         "mocha": "^5.2.0",
