+++ src/content-scripts/sidebar.ts
@@ -1,0 +1,5 @@
+import 'babel-polyfill'
+
+export async function main() {}
+
+main()
